main: rename entry_private.h to entry_p.h
