+++ main/debug.c
@@ -19,7 +19,7 @@
 #include <string.h>
 
 #include "debug.h"
-#include "entry_private.h"
+#include "entry_p.h"
 #include "options.h"
 #include "read.h"
