+++ main/entry_p.h
