+++ win32/ctags_vs2013.vcxproj
@@ -255,7 +255,7 @@
     <ClInclude Include="..\main\dependency.h" />
     <ClInclude Include="..\main\e_msoft.h" />
     <ClInclude Include="..\main\entry.h" />
-    <ClInclude Include="..\main\entry_private.h" />
+    <ClInclude Include="..\main\entry_p.h" />
     <ClInclude Include="..\main\error.h" />
     <ClInclude Include="..\main\field.h" />
     <ClInclude Include="..\main\flags.h" />
